Pulled in performance tweak from https://github.com/Arnleif/engine/commit/5f192937c260dcdc95dcaea85264eb20a5716d8b
Notes from @Arnleif:

 Hackish fix for performance loss in certain situation.

The problem, massive amount of drawcommands being sent each frame, happens if DomElement.OnDismount is called and
this.setCutoutState(false);
happens. This call seem to be to turn off this cutout thing, but in my case it was never turned on.
Anyway, it causes som webgl content to be created, just so it can set that flag to off.
And in turn that causes the barrage of drawcommands.

So this fix is hackish in the sense that it stops
context._initWebGLRenderer();
from happening, and does not adress the real problem, whatever that is..